windows: test that directory can be deleted
While testing with a third-party application on Windows, I noticed
that even after running "watchman watch-del <path>" that watchman
still had an open handle on the directory it was watching. This
prevents me from deleting that directory until I kill watchman.
This happens even though "watchman watch-list" shows that the
"watch-del" command succeeded.

It looks like the FileDescriptor object will properly close a
handle in the destructor, but the problem is that the object is
never deleted. This test demonstrates the problem.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>